Refactor zip tests for more granularity.
The single zip_test was getting too big. I want to start creating
tests for more cases, and want to reuse the comparator tooling.
