+++ tests/zip/BUILD
@@ -16,12 +16,23 @@
 load("//pkg:mappings.bzl", "pkg_attributes", "pkg_mkdirs", "pkg_mklink")
 load("//pkg:zip.bzl", "pkg_zip")
 load("//tests/util:defs.bzl", "directory", "fake_artifact")
-load("@rules_python//python:defs.bzl", "py_test")
+load("@rules_python//python:defs.bzl", "py_library", "py_test")
 load("@bazel_skylib//rules:copy_file.bzl", "copy_file")
 
 package(default_applicable_licenses = ["//:license"])
 
 licenses(["notice"])
+
+py_library(
+    name = "zip_test_lib",
+    srcs = [
+        "zip_test_lib.py",
+    ],
+    srcs_version = "PY3",
+    deps = [
+        "@bazel_tools//tools/python/runfiles",
+    ],
+)
 
 genrule(
     name = "generate_files",
@@ -204,28 +215,44 @@ py_test(
         "zip_test.py",
     ],
     data = [
-        "test_zip_empty.zip",
-        "test_zip_basic.zip",
-        "test_zip_basic_renamed",
-        # We do not actually use this file, we just want to make sure we build
-        # it with this name.
-        "test_zip_out.foo",
-        "test_zip_package_dir0.zip",
-        "test_zip_permissions.zip",
-        "test_zip_timestamp.zip",
-        "test_zip_tree.zip",
-        "test-zip-strip_prefix-empty.zip",
-        "test-zip-strip_prefix-none.zip",
-        "test-zip-strip_prefix-zipcontent.zip",
-        "test-zip-strip_prefix-dot.zip",
-
-        # these could be replaced with diff_test() rules (from skylib)
-        "test_zip_basic_timestamp_before_epoch.zip",
-        "test_zip_package_dir1.zip",
-        "test_zip_package_dir2.zip",
+        ":test-zip-strip_prefix-dot.zip",
+        ":test-zip-strip_prefix-empty.zip",
+        ":test-zip-strip_prefix-none.zip",
+        ":test-zip-strip_prefix-zipcontent.zip",
+        ":test_zip_basic.zip",
+        ":test_zip_empty.zip",
+        ":test_zip_package_dir0.zip",
+        ":test_zip_permissions.zip",
+        ":test_zip_timestamp.zip",
+        ":test_zip_tree.zip",
     ],
     python_version = "PY3",
     deps = [
+        ":zip_test_lib",
+        "@bazel_tools//tools/python/runfiles",
+    ],
+)
+
+py_test(
+    name = "zip_byte_for_byte_test",
+    srcs = [
+        "zip_byte_for_byte_test.py",
+    ],
+    data = [
+        ":test_zip_basic.zip",
+        ":test_zip_basic_renamed",
+        # We do not actually use this file, we just want to make sure we build
+        # it with this name.
+        ":test_zip_out.foo",
+        ":test_zip_package_dir0.zip",
+        # these could be replaced with diff_test() rules (from skylib)
+        ":test_zip_basic_timestamp_before_epoch.zip",
+        ":test_zip_package_dir1.zip",
+        ":test_zip_package_dir2.zip",
+    ],
+    python_version = "PY3",
+    deps = [
+        ":zip_test_lib",
         "@bazel_tools//tools/python/runfiles",
     ],
 )
